+++ src/commands/fif.js
@@ -1,4 +1,3 @@
-// Modules
 // Get modules
 const fs = require("fs");
 const chalk = require("chalk");